stub out web support for emit_otlp
