+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
     "emitter/otlp/gen",
     "emitter/otlp/test/integration",
     "emitter/otlp/test/throughput",
+    "emitter/otlp/test/web",
     "macros",
     "examples/common_patterns",
     "examples/opentelemetry/direct_otlp",
