+++ emitter/otlp/test/web/index.html
@@ -1,0 +1,40 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta content="text/html;charset=utf-8" http-equiv="Content-Type"/>
+  <title>Flare</title>
+  <script type="module">
+    import init, { setup, run } from './pkg/emit_otlp_test_integration.js';
+
+    async function initRun() {
+      await init();
+
+      setup();
+      run();
+    }
+
+    initRun();
+  </script>
+  <style>
+    body {
+      max-width: 800px;
+      margin: 0 auto;
+      padding: 0 2rem;
+      font-family: sans-serif;
+    }
+
+    p {
+      line-height: 1.2rem;
+    }
+
+    h1, h2, h3 {
+      margin-top: 2rem;
+    }
+  </style>
+</head>
+<body>
+<h1><code>emit_web</code></h1>
+<p>This sample demonstrates how we can use <code>emit</code> in <a href="https://webassembly.org/">WebAssembly</a></p>
+<p>You should see a log in your Developer Tools Console output</p>
+</body>
+</html>
