+++ emitter/otlp/test/web/run.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+wasm-pack build --target web --profiling
+docker run -it -v "$(pwd):/usr/share/nginx/html:ro" -p 8080:80 nginx
